app/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
fastapi==0.110.0
flatbuffers==24.3.7
h11==0.14.0
humanfriendly==10.0
idna==3.6
mpmath==1.3.0
numpy==1.24.4
onnx==1.15.0
onnxruntime==1.17.1
packaging==24.0
pillow==10.2.0
protobuf==4.25.3
pydantic==2.6.3
pydantic_core==2.16.3
pyreadline3==3.4.1
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
typing_extensions==4.10.0
uvicorn==0.28.0
